extract audio related code into audio utils
